pyproject.toml
[project]
name = "week7"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "gradio>=5.49.1",
    "numpy>=2.3.4",
    "openai>=2.6.1",
]
